server/requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
PyJWT==1.7.1
mysql-connector-python==8.0.25
email-validator==1.1.2